app/requirements.txt
flask
flask-swagger-ui
langchain-openai
python-dotenv